Live file download changes.
A "Live" file is one that is persisted to the `/live` directory tree.
They contain content that may change upstream based upon the
inherit concept of a "current" release. Since the contents of these
files may change without a change in the upstream name, we will
make the result more cache favorable.

When a "live" file is processed, it will now be downloaded if the theme
or plugin is not being skipped.
After the file is downloaded, it is possible to determine the .
the SHA-256 of the conten. This is used to generate a new filename.

* Add a `--hashLength` option which controls the number of
  characters in the added hash.
* A new `downloadLiveFile` function is created to handle
  download of "live" files.